requirements.txt
absl-py==2.0.0
asgiref==3.7.2
astunparse==1.6.3
attrs== 23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blosc==1.11.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
daphne==4.0.0
defusedxml==0.7.1
dill==0.3.7
Django==4.1.5
django-admin-sortable==2.2.4
django-autocomplete-light==3.9.7
django-djconfig==0.11.0
django-leaflet==0.29.0
exceptiongroup==1.1.3
flatbuffers==23.5.26
fonttools==4.43.1
fpdf2==2.7.5
gast==0.4.0
#GDAL==3.3.3
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==22.10.0
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
miceforest==5.6.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pip==23.2.1
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
selenium==4.14.0
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
trio==0.22.2
trio-websocket==0.11.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
wheel==0.41.2
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.0.0
zipp==3.17.0
zope.interface==6.1